usunalem IMU.c, teraz wzystko jest w main.c Signed-off-by: KarolZal
